requirements.txt
streamlit==1.32.0
google-generativeai==0.3.1
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
numpy==1.24.3
tqdm==4.66.1
PyPDF2==3.0.1
langchain==0.0.335
langchain-community==0.0.19
langchain-text-splitters==0.0.1
faiss-cpu==1.7.4
sentence-transformers==2.3.1 